MOVEONLY: Move logging code from util.{h,cpp} to new files.
